+++ MK4duo/src/core/commands/gcode/gcode.h
@@ -112,6 +112,8 @@
 #include "feature/m150.h"
 #include "feature/m240.h"                 // Photo Camera
 #include "feature/m600.h"                 // Advanced Pause change filament
+#include "feature/m603.h"                 // Configure filament change
+#include "feature/m701_m702.h"            // Load / Unload filament
 #include "feature/m911_m915.h"            // Set TMC2130 driver
 #include "feature/m922.h"                 // TMC DEBUG