+++ MK4duo/src/lcd/language/language_en.h
@@ -624,6 +624,12 @@
 #ifndef MSG_FILAMENT_DIAM
   #define MSG_FILAMENT_DIAM                   _UxGT("Fil. Dia.")
 #endif
+#ifndef MSG_FILAMENT_UNLOAD
+  #define MSG_FILAMENT_UNLOAD                 _UxGT("Unload mm")
+#endif
+#ifndef MSG_FILAMENT_LOAD
+  #define MSG_FILAMENT_LOAD                   _UxGT("Load mm")
+#endif
 #ifndef MSG_ADVANCE_K
   #define MSG_ADVANCE_K                       _UxGT("Advance K")
 #endif
@@ -722,6 +728,15 @@
 #endif
 #ifndef MSG_FILAMENTCHANGE
   #define MSG_FILAMENTCHANGE                  _UxGT("Change filament")
+#endif
+#ifndef MSG_FILAMENTLOAD
+  #define MSG_FILAMENTLOAD                    _UxGT("Load filament")
+#endif
+#ifndef MSG_FILAMENTUNLOAD
+  #define MSG_FILAMENTUNLOAD                  _UxGT("Unload filament")
+#endif
+#ifndef MSG_FILAMENTUNLOAD_ALL
+  #define MSG_FILAMENTUNLOAD_ALL              _UxGT("Unload All")
 #endif
 #ifndef MSG_INIT_SDCARD
   #define MSG_INIT_SDCARD                     _UxGT("Init. SD card")
@@ -1027,14 +1042,20 @@
 #ifndef MSG_DAC_EEPROM_WRITE
   #define MSG_DAC_EEPROM_WRITE                _UxGT("DAC EEPROM Write")
 #endif
-#ifndef MSG_FILAMENT_CHANGE_HEADER
-  #define MSG_FILAMENT_CHANGE_HEADER          _UxGT("PRINT PAUSED")
+#ifndef MSG_FILAMENT_CHANGE_HEADER_PAUSE
+  #define MSG_FILAMENT_CHANGE_HEADER_PAUSE    _UxGT("PRINT PAUSED")
+#endif
+#ifndef MSG_FILAMENT_CHANGE_HEADER_LOAD
+  #define MSG_FILAMENT_CHANGE_HEADER_LOAD     _UxGT("LOAD FILAMENT")
+#endif
+#ifndef MSG_FILAMENT_CHANGE_HEADER_UNLOAD
+  #define MSG_FILAMENT_CHANGE_HEADER_UNLOAD   _UxGT("UNLOAD FILAMENT")
 #endif
 #ifndef MSG_FILAMENT_CHANGE_OPTION_HEADER
   #define MSG_FILAMENT_CHANGE_OPTION_HEADER   _UxGT("RESUME OPTIONS:")
 #endif
-#ifndef MSG_FILAMENT_CHANGE_OPTION_EXTRUDE
-  #define MSG_FILAMENT_CHANGE_OPTION_EXTRUDE  _UxGT("Extrude more")
+#ifndef MSG_FILAMENT_CHANGE_OPTION_PURGE
+  #define MSG_FILAMENT_CHANGE_OPTION_PURGE    _UxGT("Purge more")
 #endif
 #ifndef MSG_FILAMENT_CHANGE_OPTION_RESUME
   #define MSG_FILAMENT_CHANGE_OPTION_RESUME   _UxGT("Resume print")
@@ -1061,11 +1082,6 @@
     #define MSG_FILAMENT_CHANGE_INIT_1          _UxGT("Wait for start")
     #define MSG_FILAMENT_CHANGE_INIT_2          _UxGT("of the filament")
     #define MSG_FILAMENT_CHANGE_INIT_3          _UxGT("change")
-  #endif
-  #ifndef MSG_FILAMENT_CHANGE_COOL_1
-    #define MSG_FILAMENT_CHANGE_COOL_1          _UxGT("Wait")
-    #define MSG_FILAMENT_CHANGE_COOL_2          _UxGT("Cool Down")
-    #define MSG_FILAMENT_CHANGE_COOL_3          _UxGT("Nozzle")
   #endif
   #ifndef MSG_FILAMENT_CHANGE_UNLOAD_1
     #define MSG_FILAMENT_CHANGE_UNLOAD_1        _UxGT("Wait for")
@@ -1094,10 +1110,10 @@
     #define MSG_FILAMENT_CHANGE_LOAD_2          _UxGT("filament load")
     #define MSG_FILAMENT_CHANGE_LOAD_3          _UxGT("")
   #endif
-  #ifndef MSG_FILAMENT_CHANGE_EXTRUDE_1
-    #define MSG_FILAMENT_CHANGE_EXTRUDE_1       _UxGT("Wait for")
-    #define MSG_FILAMENT_CHANGE_EXTRUDE_2       _UxGT("filament extrude")
-    #define MSG_FILAMENT_CHANGE_EXTRUDE_3       _UxGT("")
+  #ifndef MSG_FILAMENT_CHANGE_PURGE_1
+    #define MSG_FILAMENT_CHANGE_PURGE_1         _UxGT("Wait for")
+    #define MSG_FILAMENT_CHANGE_PURGE_2         _UxGT("filament purge")
+    #define MSG_FILAMENT_CHANGE_PURGE_3         _UxGT("")
   #endif
   #ifndef MSG_FILAMENT_CHANGE_RESUME_1
     #define MSG_FILAMENT_CHANGE_RESUME_1        _UxGT("Wait for print")
@@ -1120,8 +1136,8 @@
   #ifndef MSG_FILAMENT_CHANGE_LOAD_1
     #define MSG_FILAMENT_CHANGE_LOAD_1          _UxGT("Loading...")
   #endif
-  #ifndef MSG_FILAMENT_CHANGE_EXTRUDE_1
-    #define MSG_FILAMENT_CHANGE_EXTRUDE_1       _UxGT("Extruding...")
+  #ifndef MSG_FILAMENT_CHANGE_PURGE_1
+    #define MSG_FILAMENT_CHANGE_PURGE_1         _UxGT("Extruding...")
   #endif
   #ifndef MSG_FILAMENT_CHANGE_RESUME_1
     #define MSG_FILAMENT_CHANGE_RESUME_1        _UxGT("Resuming...")
@@ -1137,18 +1153,6 @@
 #endif
 #ifndef MSG_CONFIG
   #define MSG_CONFIG                          _UxGT("Configuration")
-#endif
-#ifndef MSG_E_BOWDEN_LENGTH
-  #define MSG_E_BOWDEN_LENGTH                 MSG_EXTRUDE _UxGT(" ") STRINGIFY(BOWDEN_LENGTH) _UxGT("mm")
-#endif
-#ifndef MSG_R_BOWDEN_LENGTH
-  #define MSG_R_BOWDEN_LENGTH                 MSG_RETRACT _UxGT(" ") STRINGIFY(BOWDEN_LENGTH) _UxGT("mm")
-#endif
-#ifndef MSG_PURGE_XMM
-  #define MSG_PURGE_XMM                       MSG_PURGE _UxGT(" ") STRINGIFY(LCD_PURGE_LENGTH) _UxGT("mm")
-#endif
-#ifndef MSG_RETRACT_XMM
-  #define MSG_RETRACT_XMM                     MSG_RETRACT _UxGT(" ") STRINGIFY(LCD_RETRACT_LENGTH) _UxGT("mm")
 #endif
 #ifndef MSG_SAVED_POS
   #define MSG_SAVED_POS                       _UxGT("Saved position")